requirements.txt
flask==3.0.0
PyGithub==2.1.1
slack-sdk==3.23.0
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0

